backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
certifi==2026.1.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fastapi==0.128.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
hydra-core==1.3.2
idna==3.11
iopath==0.1.10
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
omegaconf==2.3.0
opencv-python==4.13.0.90
packaging==26.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
portalocker==3.2.0
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
# Editable install with no version control (SAM-2==1.0)
-e /Users/varun/Documents/project_elective/backend/sam2
sam2==1.1.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.2
traitlets==5.14.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
wcwidth==0.5.3
